# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CertificationItem": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "issuer": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CertificationItem",
        "type": "object"
      },
      "EducationItem": {
        "properties": {
          "degree": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree"
          },
          "end_date": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "field_of_study": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Of Study"
          },
          "institution": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Institution",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "required": [
          "institution"
        ],
        "title": "EducationItem",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Body of every non-validation error (400, 401, 403, 404, 409, 429, 500, 503).\n\nValidation failures (422) use the standard FastAPI list form instead:\n``{\"detail\": [{\"loc\": [...], \"msg\": \"...\", \"type\": \"...\"}]}``.\nClients must treat ``detail`` as ``string | ValidationIssue[]``.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExperienceItem": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "end_date": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "start_date": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "ExperienceItem",
        "type": "object"
      },
...
      "ProfileCreate": {
        "properties": {
          "about": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "About"
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/CertificationItem"
            },
            "maxItems": 50,
            "title": "Certifications",
            "type": "array"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationItem"
            },
            "maxItems": 20,
            "title": "Education",
            "type": "array"
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/ExperienceItem"
            },
            "maxItems": 50,
            "title": "Experience",
            "type": "array"
          },
          "headline": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "profile_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectItem"
            },
            "maxItems": 50,
            "title": "Projects",
            "type": "array"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "maxItems": 200,
            "title": "Skills",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "title": "ProfileCreate",
        "type": "object"
      },
      "ProfileResponse": {
        "properties": {
          "about": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "About"
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/CertificationItem"
            },
            "title": "Certifications",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationItem"
            },
            "title": "Education",
            "type": "array"
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/ExperienceItem"
            },
            "title": "Experience",
            "type": "array"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectItem"
            },
            "title": "Projects",
            "type": "array"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "title": "Skills",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "headline",
          "about",
          "location",
          "profile_url",
          "skills",
          "experience",
          "education",
          "certifications",
          "projects",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
      "ProjectItem": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectItem",
        "type": "object"
      },
...
      "SkillItem": {
        "properties": {
          "endorsements": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endorsements"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "SkillItem",
        "type": "object"
      },
...
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
      "ValidationIssue": {
        "description": "One entry of a 422 response. Submitted values are intentionally NOT echoed.",
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Career Intelligence Copilot — Analyze your profile against live market data",
    "title": "Career Copilot API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/profile": {
      "get": {
        "description": "Get the authenticated user's profile.",
        "operationId": "get_profile_api_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The user has not saved a profile yet"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "description": "Create or update the authenticated user's profile.\n\nuser_id is derived from the auth token — never from request body.\nInput is validated and sanitized by the Pydantic schema.",
        "operationId": "upsert_profile_api_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Request body failed validation"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Upsert Profile",
        "tags": [
          "profile"
        ]
      }
    }
  }
}